Add content in the README.md file and fix linter errors
