+++ README.md
@@ -1,1 +1,70 @@
-# my-enumerable
+# CREATE AN ENUMERABLE 
+
+> This project uses ruby to implement an enumerable mixin.
+
+## Setup
+
+- Install Ruby in local environment.
+- Run a program using command line.
+
+## Prerequisites
+ - Understand Ruby basic syntax
+ - No linter errors.
+## Project requirements
+
+- use ruby to implement an enumerable mixin.
+
+- Make sure that your code is DRY.
+
+## Getting Started
+
+ To view the morse code;
+
+a) Clone the project
+
+    ```terminal
+    https://github.com/bamsi/my-enumerable.git
+    ```
+b) View the folder
+    - cd into the project directory
+
+        ```terminal
+        cd my-enumerable.git
+
+## Authors
+
+### Author 1
+
+👤 **Toussaint Saraza**
+
+- GitHub: [@sarazaToussaint](https://github.com/sarazaToussaint)
+- Twitter: [@Toussaint Saraz1](https://twitter.com/ToussaintSaraz1)
+- LinkedIn: [LinkedIn](https://www.linkedin.com/in/toussaintsaraza/)
+
+### Author 2
+
+👤 **Haji Bamsi**
+
+- GitHub: [@Bamsi](https://github.com/bamsi)
+- Twitter: [@Bamsi](https://twitter.com/bamsi)
+- LinkedIn: [@bamsi](https://linkedin.com/in/bamsi)
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](../../issues/).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Hat tip to anyone whose code was used
+- Inspiration
+- etc
+
+## 📝 License
+
+This project is [MIT](./MIT.md) licensed.
